[move] sponsors components to @/components/sponsors
